feat:rbac基本表的管理

+++ app/Http/Controllers/Admin/AdminController.php
@@ -17,8 +17,7 @@ use App\Http\Controllers\Admin\BaseController;
 class AdminController extends BaseController
 {
 
-    public function __construct()
-    {
+    public function __construct() {
         parent::__construct();
     }
 
@@ -29,7 +28,13 @@ class AdminController extends BaseController
      * 首页模板
      */
     public function index(Request $request) {
-        return view('/admin/admin/index');
+        $pageSize = 10;
+        $account = $request->get('account','');
+//        $email = $request->get('email','');
+//        $phone = $request->get('phone','');
+        $status = $request->get('status',-1);
+        list($paginate,$data) = $this->adminModel->getList($pageSize,$account,$status);
+        return view('/admin/admin/index',compact('paginate','data'));
     }
 
 
@@ -50,11 +55,9 @@ class AdminController extends BaseController
      */
     public function doAdd(AdminAddRequest $request) {
         $data = $request->post();
-
-        if(!$this->adminModel->add($data['account'],$data['password'],$data['status'],$data['isAdmin'],$data['email'],$data['phone'])) {
+        if(!$this->adminModel->add($data['account'],$data['password'],$data['status'],$data['email'],$data['phone'])) {
             return jsonPrint('001','录入失败');
         }
-
         return jsonPrint('000','录入成功');
     }
 
@@ -63,11 +66,7 @@ class AdminController extends BaseController
      * @return mixed
      * 编辑模板
      */
-    public function update(Request $request) {
-        $id = $request->get('id');
-        if(empty($id)) {
-            return redirect('/admin/admin/index');
-        }
+    public function update($id) {
         //取出旧数据
         $admin = $this->adminModel->getOne($id);
         return view('/Admin/admin/edit',compact('admin'));
@@ -84,20 +83,20 @@ class AdminController extends BaseController
         $data = $request->post();
 
         //判断数据唯一
-        if(!$this->adminModel->checkFieldIsExistsExceptIdByAdmin($data['id'],'account',$data['account'])) {
+        if($this->adminModel->checkFieldIsExistsExceptIdByAdmin($data['id'],'account',$data['account'])) {
             return jsonPrint('002','帐号已存在');
         }
 
-        if(!$this->adminModel->checkFielDIsExistsExceptIdByAdminInfo($data['id'],'email',$data['email'])) {
+        if($this->adminModel->checkFielDIsExistsExceptIdByAdminInfo($data['id'],'email',$data['email'])) {
             return jsonPrint('002','邮箱已存在');
         }
 
-        if(!$this->adminModel->checkFielDIsExistsExceptIdByAdminInfo($data['id'],'phone',$data['phone'])) {
+        if($this->adminModel->checkFielDIsExistsExceptIdByAdminInfo($data['id'],'phone',$data['phone'])) {
             return jsonPrint('002','手机号码已存在');
         }
 
 
-        if(!$this->adminModel->edit($data['id'],$data['account'],$data['password'],$data['status'],$data['isAdmin'],$data['email'],$data['phone'])) {
+        if(!$this->adminModel->edit($data['id'],$data['account'],$data['password'],$data['status'],$data['email'],$data['phone'])) {
             return jsonPrint('001','编辑失败');
         }
 
@@ -117,7 +116,9 @@ class AdminController extends BaseController
             return redirect('/admin/admin/index');
         }
 
-        $this->isOwnAdminId($id);
+        if($this->isUnDeleteAdmin($id)) {
+            return jsonPrint('001','不能删除的管理员!');
+        }
 
         if(!$this->adminModel->deleteOne($id)) {
             return jsonPrint('001','删除失败');
@@ -139,11 +140,32 @@ class AdminController extends BaseController
             return redirect('/admin/admin/index');
         }
 
-        $this->isOwnAdminId($ids);
         if(!$this->adminModel->deleteAll($ids)) {
             return jsonPrint('001','删除失败');
         }
-
         return jsonPrint('000','删除成功');
+    }
+
+
+    /**
+     * @param Request $request
+     * @return \Illuminate\Http\JsonResponse
+     * 更改管理员状态
+     */
+    public function changeStatus(Request $request) {
+        $adminId = $request->get('id');
+        if(empty($adminId)) {
+            return jsonPrint('001','非法访问!');
+        }
+
+        if($adminId  == $this->getAdminId()) {
+            return jsonPrint('001','不能禁用当前管理员!');
+        }
+
+        if(!$this->adminModel->changeStatus($adminId)) {
+            return jsonPrint('001','操作失败!');
+        }
+
+        return jsonPrint('000','操作成功!');
     }
 }
